experiments/requirements.txt
# Data extraction
clickhouse-connect>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0

# ML / autoencoder
torch>=2.0.0
scipy>=1.10.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.0

# Explainability
shap>=0.42.0

# Experiment runner
pyyaml>=6.0
